move strcut FlagcxSlice from cc to h
