+++ flagcx/core/flagcx_p2p.cc
@@ -44,8 +44,6 @@
 
 extern struct flagcxNetAdaptor flagcxNetIbP2p;
 
-struct FlagcxSlice;
-
 extern "C" flagcxResult_t flagcxP2pSliceBatch(void *sendComm,
                                               struct ibv_qp *qp, int count,
                                               FlagcxSlice **slices);
@@ -300,47 +298,6 @@ static uint64_t gNextMrId = 1;
 static std::unordered_map<uint64_t, FlagcxP2pXfer> gXferMap;
 static std::mutex gXferMutex;
 static uint64_t gNextXferId = 1;
-
-struct FlagcxTransferTask {
-  std::atomic<uint64_t> sliceCount{0};
-  std::atomic<uint64_t> doneSliceCount{0};
-  std::vector<FlagcxSlice *> sliceList;
-
-  bool isAllDone() const {
-    auto total = sliceCount.load(std::memory_order_acquire);
-    auto done = doneSliceCount.load(std::memory_order_acquire);
-    return total > 0 && done >= total;
-  }
-};
-
-enum FlagcxSliceOp : uint8_t {
-  FLAGCX_SLICE_OP_WRITE = 0,
-  FLAGCX_SLICE_OP_READ = 1,
-};
-
-struct FlagcxSlice {
-  // WRITE: local source VA; READ: local destination VA.
-  uint64_t srcVa = 0;
-  // WRITE: remote destination VA; READ: remote source VA.
-  uint64_t dstVa;
-  uint32_t length;
-  uint32_t lkey;
-  uint32_t rkey;
-  uint8_t opcode;
-  std::string peerNicPath;
-  FlagcxTransferTask *task;
-  volatile int *qpDepth;
-
-  inline void markSuccess() {
-    if (task)
-      task->doneSliceCount.fetch_add(1, std::memory_order_release);
-  }
-
-  inline void markFailed() {
-    if (task)
-      task->doneSliceCount.fetch_add(1, std::memory_order_release);
-  }
-};
 
 struct FlagcxNixlSlicePolicy {
   static constexpr bool kFurtherCut = false;
